requirements_txt.txt
torch==2.1.0
torchvision==0.16.0
opencv-python==4.8.1.78
numpy==1.24.3
pandas==2.0.3
Pillow==10.0.1
matplotlib==3.7.2
seaborn==0.12.2
scikit-learn==1.3.0
PyYAML==6.0.1
tqdm==4.66.1